Minor updates to rasterio wrappers, use crs from array
