+++ Scripts/dea_spatialtools.py
@@ -18,28 +18,270 @@ If you would like to report an issue with this script, file one on
 Github: https://github.com/GeoscienceAustralia/dea-notebooks/issues/new
 
 Functions included:
+    xr_vectorize
+    xr_rasterize
     subpixel_contours
     interpolate_2d
     contours_to_array
     largest_region
     transform_geojson_wgs_to_epsg
 
-Last modified: November 2019
+Last modified: February 2020
 
 '''
 
 # Import required packages
+import osr
+import ogr
 import collections
 import numpy as np
 import xarray as xr
 import geopandas as gpd
-import osr
-import ogr
+import rasterio.features
 import scipy.interpolate
 from scipy import ndimage as nd
 from skimage.measure import label
 from skimage.measure import find_contours
-from shapely.geometry import LineString, MultiLineString
+from shapely.geometry import LineString, MultiLineString, shape
+
+
+def xr_vectorize(da, 
+                 attribute_col='attribute', 
+                 transform=None, 
+                 crs=None, 
+                 dtype='float32',
+                 export_shp=False,
+                 **rasterio_kwargs):    
+    """
+    Vectorises a xarray.DataArray into a geopandas.GeoDataFrame.
+    
+    Parameters
+    ----------
+    da : xarray dataarray or a numpy ndarray  
+    attribute_col : str, optional
+        Name of the attribute column in the resulting geodataframe. 
+        Values of the raster object converted to polygons will be 
+        assigned to this column. Defaults to 'attribute'.
+    transform : affine.Affine object, optional
+        An affine.Affine object (e.g. `from affine import Affine; 
+        Affine(30.0, 0.0, 548040.0, 0.0, -30.0, "6886890.0) giving the 
+        affine transformation used to convert raster coordinates 
+        (e.g. [0, 0]) to geographic coordinates. If none is provided, 
+        the function will attempt to obtain an affine transformation 
+        from the xarray object (e.g. either at `da.transform` or
+        `da.geobox.transform`).
+    crs : str or CRS object, optional
+        An EPSG string giving the coordinate system of the array 
+        (e.g. 'EPSG:3577'). If none is provided, the function will 
+        attempt to extract a CRS from the xarray object's `crs` 
+        attribute.
+    dtype : str, optional
+         Data type must be one of int16, int32, uint8, uint16, 
+         or float32
+    export_shp : Boolean or string path, optional
+        To export the output vectorised features to a shapefile, supply
+        an output path (e.g. 'output_dir/output.shp'. The default is 
+        False, which will not write out a shapefile. 
+    **rasterio_kwargs : 
+        A set of keyword arguments to rasterio.features.shapes
+        Can include `mask` and `connectivity`.
+    
+    Returns
+    -------
+    gdf : Geopandas GeoDataFrame
+    
+    """
+
+    
+    # Check for a crs object
+    try:
+        crs = da.crs
+    except:
+        if crs is None:
+            raise Exception("Please add a `crs` attribute to the "
+                            "xarray.DataArray, or provide a CRS using the "
+                            "function's `crs` parameter (e.g. 'EPSG:3577')")
+            
+    # Check if transform is provided as a xarray.DataArray method.
+    # If not, require supplied Affine
+    if transform is None:
+        try:
+            # First, try to take transform info from geobox
+            transform = da.geobox.transform
+        # If no geobox
+        except:
+            try:
+                # Try getting transform from 'transform' attribute
+                transform = da.transform
+            except:
+                # If neither of those options work, raise an exception telling the 
+                # user to provide a transform
+                raise Exception("Please provide an Affine transform object using the "
+                        "`affine` parameter (e.g. `from affine import "
+                        "Affine; Affine(30.0, 0.0, 548040.0, 0.0, -30.0, "
+                        "6886890.0)`")
+    
+    # Check to see if the input is a numpy array
+    if type(da) is np.ndarray:
+        vectors = rasterio.features.shapes(source=da.astype(dtype),
+                                           transform=transform,
+                                           **rasterio_kwargs)
+    
+    else:
+        # Run the vectorizing function
+        vectors = rasterio.features.shapes(source=da.data.astype(dtype),
+                                           transform=transform,
+                                           **rasterio_kwargs)
+    
+    # Convert the generator into a list
+    vectors = list(vectors)
+    
+    # Extract the polygon coordinates and values from the list
+    polygons = [polygon for polygon, value in vectors]
+    values = [value for polygon, value in vectors]
+    
+    # Convert polygon coordinates into polygon shapes
+    polygons = [shape(polygon) for polygon in polygons]
+    
+    # Create a geopandas dataframe populated with the polygon shapes
+    gdf = gpd.GeoDataFrame(data={attribute_col: values},
+                           geometry=polygons,
+                           crs={'init': str(crs)})
+    
+    # If a file path is supplied, export a shapefile
+    if export_shp:
+        gdf.to_file(export_shp) 
+        
+    return gdf
+
+
+def xr_rasterize(gdf,
+                 da,
+                 attribute_col=False,
+                 crs=None,
+                 transform=None,
+                 name=None,
+                 x_dim='x',
+                 y_dim='y',
+                 **rasterio_kwargs):    
+    """
+    Rasterizes a geopandas.GeoDataFrame into an xarray.DataArray.
+    
+    Parameters
+    ----------
+    gdf : geopandas.GeoDataFrame
+        A geopandas.GeoDataFrame object containing the vector/shapefile
+        data you want to rasterise.
+    da : xarray.DataArray
+        The shape, coordinates, dimensions, and transform of this object 
+        are used to build the rasterized shapefile. It effectively 
+        provides a template. The attributes of this object are also 
+        appended to the output xarray.DataArray.
+    attribute_col : string, optional
+        Name of the attribute column in the geodataframe that the pixels 
+        in the raster will contain.  If set to False, output will be a 
+        boolean array of 1's and 0's.
+    crs : str, optional
+        CRS metadata to add to the output xarray. e.g. 'epsg:3577'.
+        The function will attempt get this info from the input 
+        GeoDataFrame first.
+    transform : affine.Affine object, optional
+        An affine.Affine object (e.g. `from affine import Affine; 
+        Affine(30.0, 0.0, 548040.0, 0.0, -30.0, "6886890.0) giving the 
+        affine transformation used to convert raster coordinates 
+        (e.g. [0, 0]) to geographic coordinates. If none is provided, 
+        the function will attempt to obtain an affine transformation 
+        from the xarray object (e.g. either at `da.transform` or
+        `da.geobox.transform`).
+    x_dim : str, optional
+        An optional string allowing you to override the xarray dimension 
+        used for x coordinates. Defaults to 'x'.    
+    y_dim : str, optional
+        An optional string allowing you to override the xarray dimension 
+        used for y coordinates. Defaults to 'y'.
+    **rasterio_kwargs : 
+        A set of keyword arguments to rasterio.features.rasterize
+        Can include: 'all_touched', 'merge_alg', 'dtype'.
+    
+    Returns
+    -------
+    xarr : xarray.DataArray
+    
+    """
+    
+    # Check for a crs object
+    try:
+        crs = da.crs
+    except:
+        if crs is None:
+            raise Exception("Please add a `crs` attribute to the "
+                            "xarray.DataArray, or provide a CRS using the "
+                            "function's `crs` parameter (e.g. 'EPSG:3577')")
+    
+
+    # Check if transform is provided as a xarray.DataArray method.
+    # If not, require supplied Affine
+    if transform is None:
+        try:
+            # First, try to take transform info from geobox
+            transform = da.geobox.transform
+        # If no geobox
+        except:
+            try:
+                # Try getting transform from 'transform' attribute
+                transform = da.transform
+            except:
+                # If neither of those options work, raise an exception telling the 
+                # user to provide a transform
+                raise Exception("Please provide an Affine transform object using the "
+                        "`affine` parameter (e.g. `from affine import "
+                        "Affine; Affine(30.0, 0.0, 548040.0, 0.0, -30.0, "
+                        "6886890.0)`")
+    
+    # Get the dims, coords, and output shape from da
+    da = da.squeeze()
+    y, x = da.shape
+    dims = list(da.dims)
+    xy_coords = [da[y_dim], da[x_dim]]   
+    
+    # Reproject shapefile to match CRS of raster
+    print(f'Rasterizing to match xarray.DataArray dimensions ({y}, {x}) '
+          f'and projection system/CRS (e.g. {crs})')
+    gdf_reproj = gdf.to_crs(crs=crs)
+    
+    # If an attribute column is specified, rasterise using vector 
+    # attribute values. Otherwise, rasterise into a boolean array
+    if attribute_col:
+        
+        # Use the geometry and attributes from `gdf` to create an iterable
+        shapes = zip(gdf_reproj.geometry, gdf_reproj[attribute_col])
+
+        # Convert polygons into a numpy array using attribute values
+        arr = rasterio.features.rasterize(shapes=shapes,
+                                          out_shape=(y, x),
+                                          transform=transform,
+                                          **rasterio_kwargs)
+    else:
+        # Convert polygons into a boolean numpy array 
+        arr = rasterio.features.rasterize(shapes=gdf_reproj.geometry,
+                                          out_shape=(y, x),
+                                          transform=transform,
+                                          **rasterio_kwargs)
+        
+    # Convert result to a xarray.DataArray
+    if name is not None:
+        xarr = xr.DataArray(arr,
+                           coords=xy_coords,
+                           dims=dims,
+                           attrs=da.attrs,
+                           name=name)
+    else:
+        xarr = xr.DataArray(arr,
+                   coords=xy_coords,
+                   dims=dims,
+                   attrs=da.attrs)
+        
+    return xarr
 
 
 def subpixel_contours(da,
